test: move visitor tests to integration tests directory
Move SetWellKnownProgramsVisitor tests from inline #[cfg(test)] to
tests/set_well_known_programs_visitor.rs to match the existing pattern
where all visitor tests live in the tests/ directory.
