+++ codama-korok-visitors/tests/lib.rs
@@ -8,3 +8,4 @@ mod set_accounts_visitor;
 mod set_errors_visitor;
 mod set_instructions_visitor;
 mod set_program_metadata_visitor;
+mod set_well_known_programs_visitor;
